requirements.txt
# FOREX Algorithmic Trading Bot - Dependencies

# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests and API
requests>=2.28.0

# Web framework
Flask>=2.3.0
flask-cors>=4.0.0

# Database
pymongo>=4.0.0

# Date/time utilities
python-dateutil>=2.8.0

# Excel export
openpyxl>=3.1.0

# Web scraping
cloudscraper>=1.2.60
beautifulsoup4>=4.12.0
